Cleaned up java module code
